Nowa metoda obsługi generowania dokumentów. Żadnych modeli.

+++ src/test/java/pl/openpkw/dokument/generator/webservice/ManualTest.java
@@ -10,6 +10,7 @@ import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.client.WebTarget;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
 public class ManualTest {
     
@@ -23,28 +24,32 @@ public class ManualTest {
     public void run() {
         try {
             Client client = ClientBuilder.newClient();
-            String env = DEV;
+            String env = LOCAL;
             WebTarget target = client.target("http://"+env+"/openpkw-dokument-generator/service/protocol");
 
             String form = new String(Files.readAllBytes(Paths.get(this.getClass().getResource("/form.json").toURI())));
-            System.out.println(form);
 
-            byte[] response = target.request().post(Entity.entity(form, MediaType.APPLICATION_JSON), byte[].class);
-            
-            System.out.println("Received "+response.length+" bytes.");
-            
+            Response response = target.request().post(Entity.entity(form, MediaType.APPLICATION_JSON));
+            System.out.println("Status: "+response.getStatus());
+            byte[] responseContent = (byte[]) response.readEntity(byte[].class);
+
+            if (response.getStatus() != 200) {
+                System.out.println(new String(responseContent));
+                return;
+            }
+
+            System.out.println("Received "+responseContent.length+" bytes.");
             File outputFile = new File("/temp/output.pdf");
             if (outputFile.exists()) {
                 outputFile.delete();
             }
             
             FileOutputStream outputFileStream = new FileOutputStream(outputFile);
-            outputFileStream.write(response);
+            outputFileStream.write(responseContent);
             outputFileStream.close();
             
         } catch (Exception ex) {
             throw new RuntimeException(ex);
         }
-
     }
 }
